still working on an early version

+++ Zwyssigly.ImageServer.Core/Thumbnails/Disk/DiskThumbnailRepository.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.IO;
+using System.Linq;
 using System.Threading.Tasks;
 using Zwyssigly.Functional;
 
@@ -16,16 +17,25 @@ namespace Zwyssigly.ImageServer.Thumbnails.Disk
 
         private string GeneratePath(Name gallery, ThumbnailId id)
         {
-            return Path.Combine(_options.Directory, gallery.ToString(), id.ToString());
+            var fileName = _options.FileExtensions
+                ? $"{id.Tag}.{id.FormatHint.UnwrapOrThrow().FileExtension}"
+                : id.Tag.ToString();
+
+            fileName = (_options.TagDelimiter == '/')
+                ? Path.Combine(id.ImageId.ToString(), fileName)
+                : id.ImageId.ToString() + _options.TagDelimiter + fileName;
+
+            return Path.Combine(_options.Directory, gallery.ToString(), fileName);
         }
 
         private string GeneratePath(Name gallery)
-        {
-            return Path.Combine(_options.Directory, gallery.ToString());
-        }
+            => Path.Combine(_options.Directory, gallery.ToString());
 
         public Task<Result<Unit, Error>> Delete(Name gallery, IEnumerable<ThumbnailId> ids)
         {
+            if (_options.FileExtensions && ids.Any(t => t.FormatHint.IsNone))
+                return Task.FromResult(Result.Failure<Unit, Error>(new Error(ErrorCode.ImplementationError, "Format hint is required!")));
+
             foreach (var id in ids)
                 File.Delete(GeneratePath(gallery, id));
 
@@ -34,16 +44,27 @@ namespace Zwyssigly.ImageServer.Thumbnails.Disk
 
         public Task<Result<Thumbnail, Error>> Get(Name gallery, ThumbnailId id)
         {
+            if (_options.FileExtensions && id.FormatHint.IsNone)
+                return Task.FromResult(Result.Failure<Thumbnail, Error>(new Error(ErrorCode.ImplementationError, "Format hint is required!")));
+
             var data = File.ReadAllBytes(GeneratePath(gallery, id));
             return Task.FromResult(Result.Success<Thumbnail, Error>(new Thumbnail(id, data)));
         }
 
         public Task<Result<Unit, Error>> Insert(Name gallery, IEnumerable<Thumbnail> thumbnails)
         {
+            if (_options.FileExtensions && thumbnails.Any(t => t.ThumbnailId.FormatHint.IsNone))
+                return Task.FromResult(Result.Failure<Unit, Error>(new Error(ErrorCode.ImplementationError, "Format hint is required!")));
+
             Directory.CreateDirectory(GeneratePath(gallery));
 
             foreach (var thumbnail in thumbnails)
+            {
+                if (_options.TagDelimiter == '/')
+                    Directory.CreateDirectory(Path.Combine(_options.Directory, gallery.ToString(), thumbnail.ThumbnailId.ImageId.ToString()));
+
                 File.WriteAllBytes(GeneratePath(gallery, thumbnail.ThumbnailId), thumbnail.Data);
+            }
 
             return Task.FromResult(Result.Unit<Error>());
         }
